Fix GH-18956: PHP-FPM Process Count Inconsistencies
This fixes incorrect decrement of the active number of processes. This
was done in accepting stage before incrementing which is problematic if
there are already some running processes as it decrements their number
first and result in incorrect total (lower than the actual number).

In addition it also fixes GH-14212 as accept is done just once for
keepalive connection so in this case the number of active connection
just increasing with each request.

Closes GH-19191